Software/releases/smart_desk/1.0.0/system/requirements/pip.txt
simplejson
#lxml
#cssselect
enum
gtts
pyzmq
pyowm
python-dateutil
pytz
uber_rides
mysql-connector==2.1.4
netifaces
mock
appengine
pyrebase
requests==2.11.1 #important pyrebase only works with this version
tinydb
pyaudio
#scipy
#h5py
psutil
pyspotify